requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
Flask==1.0.2
Flask-SocketIO==3.0.2
gevent==1.3.6
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.19
pyOpenSSL==18.0.0
python-engineio==2.3.1
python-socketio==2.0.0
six==1.11.0
Werkzeug==0.14.1
